add board and task verifications

+++ src/modules/task/service/implementation/TaskService.ts
@@ -5,21 +5,32 @@ import { inject, injectable } from "tsyringe";
 import { ITaskRepository } from "@modules/task/infra/database/repository/ITaskRepository";
 import { ITaskResponseDTO } from "@modules/task/dtos/ITaskResponseDTO";
 import AppError from "@shared/infra/error/AppError";
+import { IBoardRepository } from "@modules/board/infra/database/repository/IBoardRepository";
 
 @injectable()
 class TaskService implements ITaskService {
     
     constructor(
+        @inject('BoardRepository')
+        private boardRepository: IBoardRepository,
         @inject('TaskRepository')
         private taskRepository: ITaskRepository,
     ) {}
 
 
     public async create(taskDto: ITaskDTO): Promise<ITaskResponseDTO> {
+        const board = await this.boardRepository.findById(taskDto.boardId);
+
+        if (!board) throw new AppError('Board not found!', 404);
+
         return await this.taskRepository.create(taskDto);
     }
     
     public async update(taskDto: ITaskUpdateDTO): Promise<ITaskResponseDTO | undefined> {
+        const board = await this.boardRepository.findById(taskDto.boardId);
+
+        if (!board) throw new AppError('Board not found!', 404);
+
         const task = await this.taskRepository.findById(taskDto.id);
 
         if (!task) throw new AppError('Task not found!', 404);
@@ -28,6 +39,10 @@ class TaskService implements ITaskService {
     }
     
     public async delete(id: number): Promise<boolean> {
+        const task = await this.taskRepository.findById(id);
+
+        if (!task) throw new AppError('Task not found!', 404);
+
         return await this.taskRepository.delete(id);
     }
